+++ crates/plugin-api/CHANGELOG.md
@@ -8,6 +8,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Removed
+
+- The `eval_bytecode` function was deleted.
+
 ## [3.2.0] - 2025-07-24
 
 ### Added